+++ examples/light_switch/client/include/main_functions.h
@@ -13,8 +13,8 @@ static void message_handler(enum action_type action, uint8_t door_id);
 
 static void set_client_state(enum lock_state state);
 
-uint32_t send_open(simple_on_off_client_t * p_client, uint16_t destination_id);
-uint32_t send_close(simple_on_off_client_t * p_client, uint16_t destination_id);
+uint32_t send_open(simple_on_off_client_t * p_client, uint8_t destination_id);
+uint32_t send_close(simple_on_off_client_t * p_client, uint8_t destination_id);
 enum action_type set_enum (uint8_t action);
 
 #endif
